Move registry.SearchResult types to api/types/registry.
