+++ api/client/lib/image_search.go
@@ -6,7 +6,7 @@ import (
 	"net/url"
 
 	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/registry"
+	"github.com/docker/docker/api/types/registry"
 )
 
 // ImageSearch makes the docker host to search by a term in a remote registry.
